# web/contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lucen AI",
    "description": "Synthetic Identity and Deepfake Insurance-Claim Fraud Detection System",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Get Job Status",
        "description": "Get job status. Investigators see any job; claimants see only their own jobs.",
        "operationId": "get_job_status_api_v1_jobs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/artifacts/{result_id}/{name}/url": {
      "get": {
        "tags": [
          "Artifacts"
        ],
        "summary": "Get Artifact Signed Url",
        "description": "Issue a short-lived signed URL for an artifact. Requires authentication.",
        "operationId": "get_artifact_signed_url_api_v1_artifacts__result_id___name__url_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Result Id"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/artifacts/{result_id}/{name}": {
      "get": {
        "tags": [
          "Artifacts"
        ],
        "summary": "Get Artifact",
        "description": "Serve a generated artifact.\n- With valid sig+exp: verified HMAC signature.\n- Direct accessible paths: allowed for verified result artifacts.\nStrictly prevents directory traversal.",
        "operationId": "get_artifact_api_v1_artifacts__result_id___name__get",
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Result Id"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "sig",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Sig"
            }
          },
          {
            "name": "exp",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Exp"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/results/{id}/decision/draft": {
      "post": {
        "tags": [
          "Decisions",
          "Decisions"
        ],
        "summary": "Draft Decision",
        "operationId": "draft_decision_api_v1_results__id__decision_draft_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Result ID",
              "title": "Id"
            },
            "description": "Result ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionDraftRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionDraftResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/results/{id}/decision": {
      "post": {
        "tags": [
          "Decisions",
          "Decisions"
        ],
        "summary": "Submit Decision",
        "operationId": "submit_decision_api_v1_results__id__decision_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Result ID",
              "title": "Id"
            },
            "description": "Result ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DecisionDraftRequest": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action"
          },
          "reason_category": {
            "type": "string",
            "title": "Reason Category"
          },
          "investigator_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Investigator Note"
          }
        },
        "type": "object",
        "required": [
          "action",
          "reason_category"
        ],
        "title": "DecisionDraftRequest"
      },
      "DecisionDraftResponse": {
        "properties": {
          "claimant_message_en": {
            "type": "string",
            "title": "Claimant Message En"
          },
          "claimant_message_hi": {
            "type": "string",
            "title": "Claimant Message Hi"
          },
          "next_steps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Next Steps"
          },
          "slots_to_resubmit": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Slots To Resubmit"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "claimant_message_en",
          "claimant_message_hi",
          "source"
        ],
        "title": "DecisionDraftResponse"
      },
      "DecisionSubmitRequest": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action"
          },
          "reason_category": {
            "type": "string",
            "title": "Reason Category"
          },
          "claimant_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claimant Message",
            "default": ""
          },
          "claimant_message_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claimant Message En"
          },
          "claimant_message_hi": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claimant Message Hi"
          },
          "internal_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Internal Note"
          },
          "slots_to_resubmit": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Slots To Resubmit"
          }
        },
        "type": "object",
        "required": [
          "action",
          "reason_category"
        ],
        "title": "DecisionSubmitRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobStatus": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/StepStatus"
            },
            "type": "array",
            "title": "Steps"
          },
          "result_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Id"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "mode",
          "steps"
        ],
        "title": "JobStatus"
      },
...
      "StepStatus": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          }
        },
        "type": "object",
        "required": [
          "name",
          "status"
        ],
        "title": "StepStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}